Abandon VtHandler in favor of event iterators
This is a big breaking change compared to the previous API!

Based on my recent experience writing u8char and grapheme_machine I've
found that a better pattern for a streaming parser state machine like this
is to have the caller provide new input and receive a series of events in
return, which they can then handle as they wish. This avoids the complexity
of forcing a "push-based" approach onto systems that might prefer to be
event-based, while still allowing sequential processing when needed.

The approach of returning events is made more awkward here than in the
other two libraries because some of the events include borrows from inside
the VtMachine itself. Therefore the event iterator needs to be a buffer
of already-emitted events rather than having the reads from the iterator
actually modify the VtMachine state incrementally.

The approach to that in _this_ commit involves a fixed buffer that is
repeatedly consumed until everything in it becomes None, which is not
very efficient. In the next commit I'm planning to adopt a better
compromise, but I want to separate that change to the implementation
details from this very significant change to the API.