Transition now explicitly tracks position
Rather than constantly re-scanning from the start of an array of Option
we'll now keep an offset into the array and increment it for each call
to next.

This is still not ideal because this object is rather large, but this seems
like the best compromise for now. Maybe I'll have a better idea later.